move repocore client to own file, pr feedback
